requirements.txt
# SPDX-FileCopyrightText: 2025 Ruslan Mstoi <ruslan.mstoi@gmail.com>
#
# SPDX-License-Identifier: MIT

aiohttp
psycopg
pyyaml
